Updated the UI
