--- app/src/main/res/drawable/toggle_selector.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/notfilledheart" android:state_checked="false" />
-    <item android:drawable="@drawable/favouritefilledicon" android:state_checked="true" />
-</selector>
