+++ app/src/main/res/layout/activity_more_room_info2.xml
@@ -64,18 +64,33 @@
 
         <Button
             android:id="@+id/button3"
+            style="@style/CustomButtonStyle"
             android:layout_width="136dp"
             android:layout_height="56dp"
-            android:layout_marginTop="32dp"
             android:layout_marginEnd="36dp"
-            android:layout_marginBottom="32dp"
-            android:background="@color/darkblue"
-            style="@style/CustomButtonStyle"
+            android:layout_marginBottom="16dp"
             android:text="@string/BookNow"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/textView12"
-            app:layout_constraintVertical_bias="0.0" />
+            app:layout_constraintVertical_bias="0.666" />
+
+        <Button
+            android:id="@+id/button9"
+            style="@style/CustomButtonStyle"
+            android:layout_width="136dp"
+            android:layout_height="56dp"
+            android:layout_marginStart="71dp"
+            android:layout_marginTop="32dp"
+            android:layout_marginEnd="204dp"
+            android:layout_marginBottom="32dp"
+            android:text="View Ratings"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/textView12"
+            app:layout_constraintVertical_bias="1.0" />
 
         <TextView
             android:id="@+id/textView12"
@@ -93,40 +108,19 @@
 
         <Button
             android:id="@+id/button5"
-            android:layout_width="112dp"
-            android:layout_height="50dp"
-            android:layout_marginStart="52dp"
-            android:layout_marginTop="32dp"
-            android:layout_marginBottom="32dp"
-            android:background="@color/darkblue"
+            style="@style/HeartDefaultStyle"
+            android:layout_width="61dp"
+            android:layout_height="49dp"
+            android:layout_marginStart="151dp"
+            android:layout_marginEnd="7dp"
             android:backgroundTint="@color/darkblue"
-            android:text="@string/Like"
-            style="@style/CustomButtonStyle"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/textView12"
-            app:layout_constraintVertical_bias="0.0" />
-
-        <TextView
-            android:id="@+id/textView22"
-            android:layout_width="66dp"
-            android:layout_height="29dp"
-            android:layout_marginStart="16dp"
-            android:layout_marginTop="588dp"
-            android:text="Ratings"
-            android:textColor="#E91E63"
-            android:textStyle="bold"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/commentRecycler"
-            android:layout_width="415dp"
-            android:layout_height="203dp"
-            android:layout_marginStart="4dp"
-            android:layout_marginTop="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/textView22" />
+            android:elevation="10sp"
+            app:layout_constraintBottom_toBottomOf="@+id/imageView2"
+            app:layout_constraintEnd_toEndOf="@+id/imageView2"
+            app:layout_constraintHorizontal_bias="1.0"
+            app:layout_constraintStart_toStartOf="@+id/textView10"
+            app:layout_constraintTop_toTopOf="@+id/imageView2"
+            app:layout_constraintVertical_bias="0.88" />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
