+++ app/src/main/res/values-land/dimens.xml
@@ -1,0 +1,3 @@
+<resources>
+      <dimen name="fab_margin">48dp</dimen>
+   </resources>
