+++ app/src/main/res/values-w600dp/dimens.xml
@@ -1,0 +1,3 @@
+<resources>
+      <dimen name="fab_margin">48dp</dimen>
+   </resources>
