+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,3 @@
+<resources>
+      <dimen name="fab_margin">16dp</dimen>
+   </resources>
