+++ app/src/main/res/values/themes.xml
@@ -6,5 +6,5 @@
     </style>
 
     <style name="Theme.BookMyCampus" parent="Base.Theme.BookMyCampus" />
-    
+
 </resources>
